document_retrieval_system/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-text-splitters
zhipuai
geoopt
torch
numpy
PyPDF2
python-dotenv
pydantic

